requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-extensions==3.2.3
django-js-asset==3.1.2
django-richtextfield==1.6.2
django-tinymce==4.1.0
gunicorn==23.0.0
packaging==25.0
pillow==11.1.0
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg-pool==3.2.6
python-dotenv==1.1.0
pytils==0.4.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
